Add user confirmation to several commands (#37)

+++ src/capi/azext_capi/custom.py
@@ -41,7 +41,7 @@ logger = get_logger(__name__)  # pylint: disable=invalid-name
 
 
 def init_environment(cmd, prompt=True):
-    check_preqreqs(cmd, install=True)
+    check_prereqs(cmd, install=True)
     # Create a management cluster if needed
     try:
         find_management_cluster_retry(cmd.cli_ctx)
@@ -114,7 +114,7 @@ def _install_capz_components():
 
 def create_management_cluster(cmd):
     # TODO: add user confirmation
-    check_preqreqs(cmd)
+    check_prereqs(cmd)
 
     command = ["clusterctl", "init", "--infrastructure", "azure"]
     try:
@@ -124,8 +124,12 @@ def create_management_cluster(cmd):
         raise UnclassifiedUserFault("Can't locate a Kubernetes cluster") from err
 
 
-def delete_management_cluster(cmd):  # pylint: disable=unused-argument
-    # TODO: add user confirmation
+def delete_management_cluster(cmd, yes=False):  # pylint: disable=unused-argument
+    exit_if_no_management_cluster()
+    msg = 'Do you want to delete CAPZ management components from the current cluster?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return
+
     command = ["clusterctl", "delete", "--all",
                "--include-crd", "--include-namespace"]
     try:
@@ -181,11 +185,14 @@ def show_management_cluster(_cmd, yes=False):
     # TODO: echo details of the management cluster in all output formats
 
 
-def update_management_cluster(cmd):
+def update_management_cluster(cmd, yes=False):
     exit_if_no_management_cluster()
-    # Check for local prerequisites
-    check_preqreqs(cmd)
-    cmd = [
+    msg = 'Do you want to update CAPZ management components on the current cluster?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return
+    # Check for clusterctl tool
+    check_prereqs(cmd, install=yes)
+    command = [
         "clusterctl",
         "upgrade",
         "apply",
@@ -195,8 +202,8 @@ def update_management_cluster(cmd):
         "v1alpha3",
     ]
     try:
-        output = subprocess.check_output(cmd, universal_newlines=True)
-        logger.info("%s returned:\n%s", " ".join(cmd), output)
+        output = subprocess.check_output(command, universal_newlines=True)
+        logger.info("%s returned:\n%s", " ".join(command), output)
     except subprocess.CalledProcessError as err:
         raise UnclassifiedUserFault(err)
 
@@ -381,9 +388,11 @@ def get_kubeconfig(capi_name):
     return "Wrote kubeconfig file to {} ".format(filename)
 
 
-def delete_workload_cluster(cmd, capi_name):
-    # TODO: add user confirmation
+def delete_workload_cluster(cmd, capi_name, yes=False):
     exit_if_no_management_cluster()
+    msg = 'Do you want to delete this Kubernetes cluster "{}"?'.format(capi_name)
+    if not yes and not prompt_y_n(msg, default="n"):
+        return
     cmd = ["kubectl", "delete", "cluster", capi_name]
     try:
         output = subprocess.check_output(cmd, universal_newlines=True)
@@ -419,19 +428,27 @@ def update_workload_cluster(cmd, capi_name):
     raise NotImplementedError
 
 
-def check_preqreqs(cmd, install=False):
+def check_prereqs(cmd, install=False):
     # Install kubectl
     if not which("kubectl") and install:
         install_kubectl(cmd)
 
     # Install clusterctl
-    if not which("clusterctl") and install:
-        install_clusterctl(cmd)
+    if not which("kubectl") and install:
+        install_kubectl(cmd)
+    # check_clusterctl(cmd, install)
 
     # Check for required environment variables
     # TODO: remove this when AAD Pod Identity becomes the default
     for var in ["AZURE_CLIENT_ID", "AZURE_CLIENT_SECRET", "AZURE_SUBSCRIPTION_ID", "AZURE_TENANT_ID"]:
         check_environment_var(var)
+
+
+def check_clusterctl(cmd, install=False):
+    if not which("clusterctl"):
+        logger.warning("clusterctl was not found.")
+        if install or prompt_y_n("Download and install clusterctl?", default="n"):
+            install_clusterctl(cmd)
 
 
 def check_environment_var(var):
